auth added

+++ src/main/java/com/mp/order/OrderController.java
@@ -1,10 +1,12 @@
 package com.mp.order;
 
 import com.mp.PizzaServiceImpl;
+import com.mp.auth.AuthServiceImpl;
 import com.mp.cart.Cart;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -18,9 +20,7 @@ public class OrderController {
     @RequestMapping(value = "/all", method = RequestMethod.GET)
     public @ResponseBody
     List<Order> getOrders() {
-        List<Order> orders = orderService.findAll();
-        System.out.println(orders);
-        return orders;
+        return orderService.findAll();
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
@@ -32,20 +32,40 @@ public class OrderController {
     @RequestMapping(value = "/make", method = RequestMethod.GET)
     @ResponseBody
     public String makeOrder(){
-        order();
-        return "Now we handle your order.";
+        return order();
     }
 
     @DeleteMapping(value = "/delete/{id}")
     @ResponseBody
-    public void deletePizza(@PathVariable Long id){
-        orderService.deleteById(id);
+    public void deleteOrder(@PathVariable Long id){
+        String username = AuthServiceImpl.getCurrentUser();
+        if (!username.isEmpty()) {
+            Optional<Order> order = orderService.getById(id);
+            if (order.isPresent()) {
+                if (order.get().getUser().equals(username))
+                    orderService.deleteById(id);
+            }
+        }
     }
 
-    private void order(){
-        Order newOrder = new Order();
-        newOrder.setPizzaIds(new Cart().getAllFromCart());
-        System.out.println("A_A_A_A__A_A_A_A_A_A_A__A_A__    " + newOrder.getId() + "\n\n\n");
-        orderService.save(newOrder);
+    private String order(){
+        String username = AuthServiceImpl.getCurrentUser();
+        if (!username.isEmpty()) {
+            Order newOrder = new Order();
+            newOrder.setPizzaIds(new Cart().getAllFromCart());
+            newOrder.setUser(username);
+            orderService.save(newOrder);
+            return "Now we handle your order.";
+        }
+        else return "You are not authorized";
+    }
+
+    @RequestMapping(value = "/my", method = RequestMethod.GET)
+    public @ResponseBody
+    List<Order> getOrdersOfUser() {
+        String username = AuthServiceImpl.getCurrentUser();
+        if (!username.isEmpty())
+            return orderService.getAllWithUsername(AuthServiceImpl.getCurrentUser());
+        else return null;
     }
 }
